refactor(probe): extract shared case and profile primitives
